Rename SinglePassCompiler to ConcreteSinglePassCompiler
