+++ Source/DafnyCore/Compilers/Synthesizer-Csharp.cs
@@ -68,7 +68,7 @@ public class CsharpSynthesizer {
   /// }
   /// </summary>
   public ConcreteSyntaxTree SynthesizeMethod(Method method,
-    List<SinglePassCompiler.TypeArgumentInstantiation> typeArgs, bool createBody,
+    List<ConcreteSinglePassCompiler.TypeArgumentInstantiation> typeArgs, bool createBody,
     ConcreteSyntaxTree wr, bool forBodyInheritance, bool lookasideBody) {
 
     lastSynthesizedMethod = method;
@@ -78,7 +78,7 @@ public class CsharpSynthesizer {
                          compiler.NeedsCustomReceiver(method);
     var keywords = compiler.Keywords(true, true);
     var returnType = compiler.GetTargetReturnTypeReplacement(method, wr);
-    var typeParameters = compiler.TypeParameters(SinglePassCompiler.TypeArgumentInstantiation.
+    var typeParameters = compiler.TypeParameters(ConcreteSinglePassCompiler.TypeArgumentInstantiation.
       ToFormals(compiler.ForTypeParameters(typeArgs, method, lookasideBody)));
     var parameters = compiler
       .GetMethodParameters(method, typeArgs, lookasideBody, customReceiver, returnType);
